refact: zoom 로직 hooks, lib 분리
