+++ src/features/diagram/lib/grid.ts
@@ -17,10 +17,6 @@ export const DEFAULT_GRID_DIMENSIONS: GridDimensions = {
 	rows: GRID_ROW_COUNT,
 };
 export const CELL_SIZE = 106;
-export const MIN_GRID_ZOOM = 0.2;
-export const MAX_GRID_ZOOM = 1;
-export const GRID_ZOOM_BUTTON_CELL_STEP = 2;
-export const GRID_ZOOM_WHEEL_CELL_STEP = 1;
 
 export function getGridPixelSize(
 	gridDimensions: GridDimensions = DEFAULT_GRID_DIMENSIONS,
@@ -51,250 +47,6 @@ export function getNodesOutsideGrid(
 	gridDimensions: GridDimensions = DEFAULT_GRID_DIMENSIONS,
 ): DiagramNode[] {
 	return nodes.filter((node) => !isNodeFullyInsideGrid(node, gridDimensions));
-}
-
-export function getResponsiveGridZoom(containerWidth: number): number {
-	return getGridZoomForVisibleCells(
-		containerWidth,
-		getResponsiveGridVisibleCellCount(containerWidth),
-	);
-}
-
-export function clampGridZoom(zoom: number): number {
-	return Math.min(Math.max(zoom, MIN_GRID_ZOOM), MAX_GRID_ZOOM);
-}
-
-export function getVisibleCellCountBounds(
-	containerWidth: number,
-	cellCount = GRID_COLUMN_COUNT,
-): { min: number; max: number } {
-	if (containerWidth <= 0) {
-		return { min: cellCount, max: cellCount };
-	}
-
-	const minVisibleCellCount = Math.min(
-		cellCount,
-		Math.max(1, Math.ceil(containerWidth / (CELL_SIZE * MAX_GRID_ZOOM))),
-	);
-	const maxVisibleCellCount = Math.min(
-		cellCount,
-		Math.max(
-			minVisibleCellCount,
-			Math.floor(containerWidth / (CELL_SIZE * MIN_GRID_ZOOM)),
-		),
-	);
-
-	return {
-		min: minVisibleCellCount,
-		max: maxVisibleCellCount,
-	};
-}
-
-export function getClampedVisibleCellCount(
-	containerWidth: number,
-	visibleCellCount: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	const bounds = getVisibleCellCountBounds(containerWidth, cellCount);
-	const roundedVisibleCellCount = Math.round(visibleCellCount);
-
-	return Math.min(Math.max(roundedVisibleCellCount, bounds.min), bounds.max);
-}
-
-export function getResponsiveGridVisibleCellCount(
-	containerWidth: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	return getVisibleCellCountBounds(containerWidth, cellCount).max;
-}
-
-export function getVisibleCellCount(
-	containerWidth: number,
-	zoom: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	if (containerWidth <= 0) return cellCount;
-
-	return getClampedVisibleCellCount(
-		containerWidth,
-		containerWidth / (CELL_SIZE * clampGridZoom(zoom)),
-		cellCount,
-	);
-}
-
-export function getGridZoomForVisibleCells(
-	containerWidth: number,
-	visibleCellCount: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	if (containerWidth <= 0) return MAX_GRID_ZOOM;
-
-	const clampedVisibleCellCount = getClampedVisibleCellCount(
-		containerWidth,
-		visibleCellCount,
-		cellCount,
-	);
-
-	return clampGridZoom(containerWidth / (CELL_SIZE * clampedVisibleCellCount));
-}
-
-export function getNextGridZoom(
-	currentZoom: number,
-	containerWidth: number,
-	visibleCellDelta: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	const currentVisibleCellCount = getVisibleCellCount(
-		containerWidth,
-		currentZoom,
-		cellCount,
-	);
-	return getGridZoomForVisibleCells(
-		containerWidth,
-		currentVisibleCellCount + visibleCellDelta,
-		cellCount,
-	);
-}
-
-export function getNextGridVisibleCellCount(
-	currentVisibleCellCount: number,
-	containerWidth: number,
-	visibleCellDelta: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	return getClampedVisibleCellCount(
-		containerWidth,
-		currentVisibleCellCount + visibleCellDelta,
-		cellCount,
-	);
-}
-
-export function getWheelGridZoom(
-	currentZoom: number,
-	deltaY: number,
-	containerWidth: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	if (deltaY === 0) return currentZoom;
-
-	const visibleCellDelta = deltaY < 0 ? -1 : 1;
-	return getNextGridZoom(
-		currentZoom,
-		containerWidth,
-		visibleCellDelta * GRID_ZOOM_WHEEL_CELL_STEP,
-		cellCount,
-	);
-}
-
-export function getWheelGridVisibleCellCount(
-	currentVisibleCellCount: number,
-	deltaY: number,
-	containerWidth: number,
-	cellCount = GRID_COLUMN_COUNT,
-): number {
-	if (deltaY === 0) return currentVisibleCellCount;
-
-	const visibleCellDelta = deltaY < 0 ? -1 : 1;
-	return getNextGridVisibleCellCount(
-		currentVisibleCellCount,
-		containerWidth,
-		visibleCellDelta * GRID_ZOOM_WHEEL_CELL_STEP,
-		cellCount,
-	);
-}
-
-export function isGridZoomWheelEvent(event: {
-	metaKey: boolean;
-	ctrlKey: boolean;
-}): boolean {
-	return event.metaKey || event.ctrlKey;
-}
-
-export function getCenteredScrollOffset(
-	center: number,
-	zoom: number,
-	viewportSize: number,
-): number {
-	return center * zoom - viewportSize / 2;
-}
-
-export function getAnchoredScrollOffset(
-	flowPosition: number,
-	anchorPosition: number,
-	zoom: number,
-): number {
-	return flowPosition * zoom - anchorPosition;
-}
-
-function clampScrollOffset(
-	scrollOffset: number,
-	zoom: number,
-	viewportSize: number,
-	axisCellCount = GRID_COLUMN_COUNT,
-): number {
-	const maxScrollOffset = Math.max(
-		0,
-		axisCellCount * CELL_SIZE * zoom - viewportSize,
-	);
-
-	return Math.min(Math.max(scrollOffset, 0), maxScrollOffset);
-}
-
-function snapScrollOffsetToCell(scrollOffset: number, zoom: number): number {
-	const renderedCellSize = CELL_SIZE * zoom;
-
-	if (renderedCellSize <= 0) return scrollOffset;
-
-	return Math.round(scrollOffset / renderedCellSize) * renderedCellSize;
-}
-
-export function getCellAlignedCenteredScrollOffset(
-	center: number,
-	zoom: number,
-	viewportSize: number,
-	axisCellCount = GRID_COLUMN_COUNT,
-): number {
-	const rawScrollOffset = getCenteredScrollOffset(center, zoom, viewportSize);
-	const clampedScrollOffset = clampScrollOffset(
-		rawScrollOffset,
-		zoom,
-		viewportSize,
-		axisCellCount,
-	);
-	const snappedScrollOffset = snapScrollOffsetToCell(clampedScrollOffset, zoom);
-	return clampScrollOffset(
-		snappedScrollOffset,
-		zoom,
-		viewportSize,
-		axisCellCount,
-	);
-}
-
-export function getCellAlignedAnchoredScrollOffset(
-	flowPosition: number,
-	anchorPosition: number,
-	zoom: number,
-	viewportSize: number,
-	axisCellCount = GRID_COLUMN_COUNT,
-): number {
-	const rawScrollOffset = getAnchoredScrollOffset(
-		flowPosition,
-		anchorPosition,
-		zoom,
-	);
-	const clampedScrollOffset = clampScrollOffset(
-		rawScrollOffset,
-		zoom,
-		viewportSize,
-		axisCellCount,
-	);
-	const snappedScrollOffset = snapScrollOffsetToCell(clampedScrollOffset, zoom);
-	return clampScrollOffset(
-		snappedScrollOffset,
-		zoom,
-		viewportSize,
-		axisCellCount,
-	);
 }
 
 export function toCellKey(cell: CellCoord): string {
